Add option for waiting after typing
There's now an optional function argument to the run finction that will
be awaited after the element is done typing. By default, this will be
the elements wait attribute.

The new resumed event and resume function can be used together with the
above described feature to pause an element until it is manually
resumed.